submissions/salman_guliwala/requirements.txt
transformers
torch
torchvision
gradio
pillow
sentence-transformers
faiss-cpu
